requirements.txt
# OCR + image processing
opencv-python>=4.8,<5
pytesseract>=0.3.10
Pillow>=10.0
numpy>=1.24

# Translation + language detection
deep-translator>=1.11
langdetect>=1.0.9

# Web framework
fastapi>=0.110
uvicorn[standard]>=0.29

# Templating
jinja2>=3.1
python-multipart>=0.0.9

httpx               # Claude Vision API calls
anthropic           # optional — or use httpx directly as coded
python-bidi         # RTL support in overlay
arabic-reshaper     # Arabic glyph reshaping
lingua-language-detector  # optional stronger langdetect replacement
